+++ src/output/mod.rs
@@ -1,7 +1,9 @@
 mod csv;
+mod html;
 pub mod json;
 mod markdown;
 
 pub use csv::{print_csv, write_csv};
+pub use html::{print_html, write_html};
 pub use json::{print_json, write_json};
 pub use markdown::{print_markdown, write_markdown};